+++ src/extension.ts
@@ -1,7 +1,6 @@
 import * as Config from './config/Config';
 import { window, ExtensionContext, commands } from 'vscode';
 import WebSearchProviderDefinition from './search/WebSearchProviderDefinition';
-import { EXTENSION_NAME } from './config/Config';
 import WebSearchProvider from './search/WebSearchProvider';
 import InputBox from './inputbox/InputBox';
 
@@ -12,7 +11,7 @@ import InputBox from './inputbox/InputBox';
  */
 export function activate(context: ExtensionContext) {
     if (!Config.isValid()) {
-        window.showErrorMessage("Web Search: Invalid configuration");
+        window.showErrorMessage("Web Search: Invalid configuration. The configuration has been updated recently: please follow the new guideline.");
     }
 
     const searchProviderDefinitions = Config.getSearchProvidersFromConfig();
@@ -36,13 +35,13 @@ export function deactivate() { }
 function registerSearchCommand(context: ExtensionContext, definition: WebSearchProviderDefinition): void {
     const commandName = definition.name;
 
-    let disposable = commands.registerCommand(`${EXTENSION_NAME}.${commandName}`, () => {
+    let disposable = commands.registerCommand(`${Config.EXTENSION_NAME}.${commandName}`, () => {
         // The code you place here will be executed every time your command is executed
         const searchProvider = getSearchProvider(definition);
         const text = getSelectedText();
-        const inputBox = new InputBox(definition.name);
 
         if (!text && Config.isInputBoxActive()) {
+            const inputBox = new InputBox(definition.name);
             inputBox.show().then((text) => {
                 searchProvider.open(text);
             });
@@ -55,19 +54,10 @@ function registerSearchCommand(context: ExtensionContext, definition: WebSearchP
 }
 
 /**
- * Retrieves a search provider object instance from the configured search or returns the default one
+ * Retrieves a search provider object instance
  * @param definition 
  */
 function getSearchProvider(definition: WebSearchProviderDefinition): WebSearchProvider {
-    if (definition.className) {
-        import(`./search/${definition.className}`)
-            .then(searchProvider => {
-                return new searchProvider.default(definition);
-            }).catch(error => {
-                window.showErrorMessage('WebSearch Extension provider class not found ' + error);
-            });
-    }
-
     return new WebSearchProvider(definition);
 }