+++ CHANGELOG.md
@@ -2,8 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](http://keepachangelog.com/) 
+The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
+
+## 1.3.0 - 2020-10-04
+Better configuration and use open instead of opn
+
+### Added
+- Indvidual search provider configuration
+- Use open package
 
 ## 1.2.0 - 2019-04-07
 First version